edit BigDecimal

+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperSvBalGoodsRepository.java
@@ -1,13 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperSvBalGoods;
 
-public interface TaPaperSvBalGoodsRepository
-    extends CommonJpaCrudRepository<TaPaperSvBalGoods, BigDecimal>
-{
-
+public interface TaPaperSvBalGoodsRepository extends CommonJpaCrudRepository<TaPaperSvBalGoods, Long> {
 
 }
