+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperSvQtyRepository.java
@@ -1,13 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperSvQty;
 
-public interface TaPaperSvQtyRepository
-    extends CommonJpaCrudRepository<TaPaperSvQty, BigDecimal>
-{
-
+public interface TaPaperSvQtyRepository extends CommonJpaCrudRepository<TaPaperSvQty, Long> {
 
 }
